cctesting-main/careercorner2/requirements.txt
streamlit>=1.40.0
google-genai==1.41.0
langfuse>=3.11.1
python-dotenv>=1.0.1
reportlab>=4.2.2
werkzeug>=3.0.4
folium>=0.15.0
streamlit-folium>=0.25.0

